topology/be/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-api-python-client==2.117.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-auth==2.27.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
websockets==12.0
redis==5.0.1
httpx==0.25.2
pillow==10.1.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
aiofiles==23.2.1
email-validator==2.1.0